move fft test from mri to test_physics
