+++ deepinv/physics/mri.py
@@ -210,7 +210,7 @@ class MultiCoilMRI(MRIMixin, LinearPhysics):
 
     .. note::
 
-        You can also simulate birdcage coil sensitivity maps by passing instead an integer to ``coil_maps``
+        You can also simulate basic `birdcage coil sensitivity maps <https://mriquestions.com/birdcage-coil.html>` by passing instead an integer to ``coil_maps``
         using ``MultiCoilMRI(coil_maps=N, img_size=x.shape)`` (note this requires ``sigpy``).
 
     :param torch.Tensor mask: binary sampling mask which should have shape (H,W), (C,H,W), (B,C,H,W), or (B,C,...,H,W). If None, generate blank mask with ``img_size``.
@@ -538,146 +538,3 @@ class SequentialMRI(DynamicMRI):
             return self.to_static().A_adjoint(
                 self.average(y, mask), mask=self.average(mask), **kwargs
             )
-
-
-# Test that our FFT is the same as FastMRI FFT implementation
-# reference: https://github.com/facebookresearch/fastMRI/blob/main/fastmri/fftc.py
-
-if __name__ == "__main__":
-
-    def fft2c_new(data: torch.Tensor, norm: str = "ortho") -> torch.Tensor:
-        r"""
-        Apply centered 2 dimensional Fast Fourier Transform.
-        :param torch.Tensor data: Complex valued input data containing at least 3 dimensions:
-            dimensions -3 & -2 are spatial dimensions and dimension -1 has size
-            2. All other dimensions are assumed to be batch dimensions.
-        :param bool norm: Normalization mode. See :meth:`torch.fft.fft`.
-        :return: (torch.tensor) the FFT of the input.
-        """
-        if not data.shape[-1] == 2:
-            raise ValueError("Tensor does not have separate complex dim.")
-
-        data = ifftshift(data, dim=[-3, -2])
-        data = torch.view_as_real(
-            torch.fft.fftn(  # type: ignore
-                torch.view_as_complex(data), dim=(-2, -1), norm=norm
-            )
-        )
-        data = fftshift(data, dim=[-3, -2])
-
-        return data
-
-    def ifft2c_new(data: torch.Tensor, norm: str = "ortho") -> torch.Tensor:
-        """
-        Apply centered 2-dimensional Inverse Fast Fourier Transform.
-        Args:
-            data: Complex valued input data containing at least 3 dimensions:
-                dimensions -3 & -2 are spatial dimensions and dimension -1 has size
-                2. All other dimensions are assumed to be batch dimensions.
-            norm: Normalization mode. See ``torch.fft.ifft``.
-        Returns:
-            The IFFT of the input.
-        """
-        if not data.shape[-1] == 2:
-            raise ValueError("Tensor does not have separate complex dim.")
-
-        data = ifftshift(data, dim=[-3, -2])
-        data = torch.view_as_real(
-            torch.fft.ifftn(  # type: ignore
-                torch.view_as_complex(data), dim=(-2, -1), norm=norm
-            )
-        )
-        data = fftshift(data, dim=[-3, -2])
-
-        return data
-
-    def roll_one_dim(x: torch.Tensor, shift: int, dim: int) -> torch.Tensor:
-        """
-        Similar to roll but for only one dim.
-        Args:
-            x: A PyTorch tensor.
-            shift: Amount to roll.
-            dim: Which dimension to roll.
-        Returns:
-            Rolled version of x.
-        """
-        shift = shift % x.size(dim)
-        if shift == 0:
-            return x
-
-        left = x.narrow(dim, 0, x.size(dim) - shift)
-        right = x.narrow(dim, x.size(dim) - shift, shift)
-
-        return torch.cat((right, left), dim=dim)
-
-    def roll(x: torch.Tensor, shift: List[int], dim: List[int]) -> torch.Tensor:
-        """
-        Similar to np.roll but applies to PyTorch Tensors.
-        Args:
-            x: A PyTorch tensor.
-            shift: Amount to roll.
-            dim: Which dimension to roll.
-        Returns:
-            Rolled version of x.
-        """
-        if len(shift) != len(dim):
-            raise ValueError("len(shift) must match len(dim)")
-
-        for s, d in zip(shift, dim):
-            x = roll_one_dim(x, s, d)
-
-        return x
-
-    def fftshift(x: torch.Tensor, dim: Optional[List[int]] = None) -> torch.Tensor:
-        """
-        Similar to np.fft.fftshift but applies to PyTorch Tensors
-        Args:
-            x: A PyTorch tensor.
-            dim: Which dimension to fftshift.
-        Returns:
-            fftshifted version of x.
-        """
-        if dim is None:
-            # this weird code is necessary for toch.jit.script typing
-            dim = [0] * (x.dim())
-            for i in range(1, x.dim()):
-                dim[i] = i
-
-        # also necessary for torch.jit.script
-        shift = [0] * len(dim)
-        for i, dim_num in enumerate(dim):
-            shift[i] = x.shape[dim_num] // 2
-
-        return roll(x, shift, dim)
-
-    def ifftshift(x: torch.Tensor, dim: Optional[List[int]] = None) -> torch.Tensor:
-        """
-        Similar to np.fft.ifftshift but applies to PyTorch Tensors
-        Args:
-            x: A PyTorch tensor.
-            dim: Which dimension to ifftshift.
-        Returns:
-            ifftshifted version of x.
-        """
-        if dim is None:
-            # this weird code is necessary for toch.jit.script typing
-            dim = [0] * (x.dim())
-            for i in range(1, x.dim()):
-                dim[i] = i
-
-        # also necessary for torch.jit.script
-        shift = [0] * len(dim)
-        for i, dim_num in enumerate(dim):
-            shift[i] = (x.shape[dim_num] + 1) // 2
-
-        return roll(x, shift, dim)
-
-    x = torch.randn(4, 2, 16, 8)
-
-    # Our FFT
-    xf1 = MRIMixin.from_torch_complex(MRIMixin.fft(MRIMixin.to_torch_complex(x)))
-
-    # FastMRI FFT
-    xf2 = fft2c_new(x.moveaxis(1, -1).contiguous()).moveaxis(-1, 1)
-
-    assert torch.all(xf1 == xf2)
